+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Since 0.4.2
 * [#94] Added delete link to Lexeme show page
 * [#204] Fixed an issue where most wanted wasn't updating
+* [#175] Will now strip leading and trailing whitespace
 
 ## Since 0.4.1
 * [#89] Lexeme form now arranges etymology horizontally instead of by nesting